requirements.txt
selenium
behave
allure-behave
allure-python-commons
PyYAML
requests
flake8
pylint
black
mypy
radon
bandit
pre-commit
